SEMBRA CHE CON K COSTANTE FUNZIONI A 100

+++ Somma.c
@@ -9,7 +9,7 @@ int somma(char *prova){
 int main(void){
     char stringa[5];
     int k = 10;
-    
+
     /*
     int prova[123*k];
 
@@ -29,7 +29,7 @@ int main(void){
 
     while(fgets(stringa,5*k,stdin)!=NULL){
         printf("%s %d\n", stringa, stringa[0]);
-        prova[stringa[0]] ++;
+        prova[(int) stringa[0]] ++;
     }
 
     for(int i=0;i<123;i++){
